Not finished version of register page for lender, has some spacing issues
