+++ frontend/src/components/routing/AppRoutes.tsx
@@ -3,9 +3,9 @@ import { useAuth } from '../../contexts/AuthContext';
 import HomePage from '../../pages/HomePage';
 import LoginPage from '../../pages/LoginPage';
 import ProfilePage from '../../pages/ProfilePage';
-import CustomerRegisterPage from '../../pages/CustomerRegisterPage';
+import CustomerRegisterPage from '../../pages/LenderRegisterPage';
 import AppRoute from './AppRoute';
-import CompanyRegisterPage from '../../pages/CompanyRegisterPage';
+import CompanyRegisterPage from '../../pages/BorrowerRegisterPage';
 import RegistrationApprovePage from '../../pages/RegistrationApprovePage';
 import FAQPage from '../../pages/FAQPage';
 import ContactInfoPage from '../../pages/ContactInfoPage';
