+++ frontend/src/pages/BorrowerRegisterPage.tsx
